+++ Changelog.md
@@ -17,6 +17,7 @@
 - Removed the implicit `status=open` filter from '/manage/order' if no explicit status query was given  
 - Improved the UI
 - Fixed the link to master products on the order item list
+- Added the JSON variant of product index responses
 - Added optional pricelist plugin support to channel form
 - Added the `currency` field to the channels form (VFW 4 feature)
 - Added the `is_hidden` field to property screens